Refactor multimodal embedding code
