+++ comps/embeddings/native/multimodal/bridgetower/dependency/Dockerfile
@@ -13,13 +13,13 @@ ENV PYTHONPATH=/home/user:/usr/lib/habanalabs/:/optimum-habana
 COPY --chown=user comps /home/user/comps
 
 RUN pip install --no-cache-dir --upgrade pip && \
-    pip install --no-cache-dir -r /home/user/comps/embeddings/multimodal/multimodal_langchain/requirements.txt
+    pip install --no-cache-dir -r /home/user/comps/embeddings/native/multimodal/bridgetower/wrapper/requirements.txt
 
 ENV PYTHONPATH=$PYTHONPATH:/home/user
 
 ARG EMBEDDER_PORT=8080
 ENV PORT=$EMBEDDER_PORT
 
-WORKDIR /home/user/comps/embeddings/multimodal/bridgetower
+WORKDIR /home/user/comps/embeddings/native/multimodal/bridgetower/dependency/
 
 ENTRYPOINT ["python", "bridgetower_server.py", "--device", "cpu"]
