+++ comps/embeddings/native/multimodal/bridgetower/dependency/Dockerfile.intel_hpu
@@ -17,7 +17,7 @@ COPY --chown=user comps /home/user/comps
 
 # Install requirements and optimum habana
 RUN pip install --no-cache-dir --upgrade pip && \
-    pip install --no-cache-dir -r /home/user/comps/embeddings/multimodal/multimodal_langchain/requirements.txt && \
+    pip install --no-cache-dir -r /home/user/comps/embeddings/native/multimodal/bridgetower/wrapper/requirements.txt && \
     pip install --no-cache-dir optimum[habana]
 
 ENV PYTHONPATH=$PYTHONPATH:/home/user
@@ -25,5 +25,5 @@ ENV PYTHONPATH=$PYTHONPATH:/home/user
 ARG EMBEDDER_PORT=8080
 ENV PORT=$EMBEDDER_PORT
 
-WORKDIR /home/user/comps/embeddings/multimodal/bridgetower
+WORKDIR /home/user/comps/embeddings/native/multimodal/bridgetower/dependency/
 ENTRYPOINT ["python", "bridgetower_server.py", "--device", "hpu"]
