+++ comps/embeddings/native/multimodal/bridgetower/dependency/docker_compose_bridgetower_embedding_intel_cpu.yaml
