+++ comps/embeddings/native/multimodal/bridgetower/wrapper/docker_compose_multimodal_embedding.yaml
