+++ comps/embeddings/native/multimodal/bridgetower/wrapper/local_mm_embedding.py
@@ -14,7 +14,7 @@ from comps import (
     opea_microservices,
     register_microservice,
 )
-from comps.embeddings.multimodal.bridgetower import BridgeTowerEmbedding
+from comps.embeddings.native.multimodal.bridgetower import BridgeTowerEmbedding
 
 logger = CustomLogger("local_multimodal_embedding")
 logflag = os.getenv("LOGFLAG", False)
