+++ comps/embeddings/native/multimodal/bridgetower/wrapper/requirements.txt
