+++ comps/embeddings/native/multimodal/clip/Dockerfile
@@ -19,10 +19,10 @@ COPY comps /home/user/comps
 
 RUN pip install --no-cache-dir --upgrade pip setuptools && \
     if [ ${ARCH} = "cpu" ]; then pip install --no-cache-dir torch torchvision --index-url https://download.pytorch.org/whl/cpu; fi && \
-    pip install --no-cache-dir -r /home/user/comps/embeddings/multimodal_clip/requirements.txt
+    pip install --no-cache-dir -r /home/user/comps/embeddings/native/multimodal/clip/requirements.txt
 
 ENV PYTHONPATH=$PYTHONPATH:/home/user
 
-WORKDIR /home/user/comps/embeddings/multimodal_clip
+WORKDIR /home/user/comps/embeddings/native/multimodal/clip/
 
 ENTRYPOINT ["python", "embedding_multimodal.py"]
