+++ comps/embeddings/native/multimodal/clip/__init__.py
