+++ comps/embeddings/native/multimodal/clip/docker_compose_embedding.yaml
