+++ comps/embeddings/native/multimodal/clip/embeddings_clip.py
