+++ comps/embeddings/native/multimodal/clip/requirements.txt
