+++ tests/embeddings/test_embeddings_multimodal_bridgetower_on_intel_hpu.sh
@@ -16,20 +16,20 @@ unset http_proxy
 function build_mmei_docker_images() {
     cd $WORKPATH
     echo $(pwd)
-    docker build --no-cache -t opea/embedding-multimodal-bridgetower:latest --build-arg EMBEDDER_PORT=$EMBEDDER_PORT --build-arg https_proxy=$https_proxy --build-arg http_proxy=$http_proxy -f comps/embeddings/multimodal/bridgetower/Dockerfile.intel_hpu .
+    docker build --no-cache -t opea/embedding-multimodal-bridgetower-hpu:latest --build-arg EMBEDDER_PORT=$EMBEDDER_PORT --build-arg https_proxy=$https_proxy --build-arg http_proxy=$http_proxy -f comps/embeddings/native/multimodal/bridgetower/dependency/Dockerfile.intel_hpu .
 
     if [ $? -ne 0 ]; then
-        echo "opea/embedding-multimodal-bridgetower built fail"
+        echo "opea/embedding-multimodal-bridgetower-hpu built fail"
         exit 1
     else
-        echo "opea/embedding-multimodal-bridgetower built successful"
+        echo "opea/embedding-multimodal-bridgetower-hpu built successful"
     fi
 }
 
 function build_embedding_service_images() {
     cd $WORKPATH
     echo $(pwd)
-    docker build --no-cache -t opea/embedding-multimodal:latest --build-arg https_proxy=$https_proxy --build-arg http_proxy=$http_proxy -f comps/embeddings/multimodal/multimodal_langchain/Dockerfile .
+    docker build --no-cache -t opea/embedding-multimodal:latest --build-arg https_proxy=$https_proxy --build-arg http_proxy=$http_proxy -f comps/embeddings/native/multimodal/bridgetower/wrapper/Dockerfile .
 
     if [ $? -ne 0 ]; then
         echo "opea/embedding-multimodal built fail"
@@ -46,10 +46,10 @@ function build_docker_images() {
 
 function start_service() {
     cd $WORKPATH
-    cd comps/embeddings/multimodal/bridgetower/
-    docker compose -f docker_compose_bridgetower_embedding_endpoint.yaml up -d
+    cd comps/embeddings/native/multimodal/bridgetower/dependency
+    docker compose -f docker_compose_bridgetower_embedding_intel_hpu.yaml up -d
     cd $WORKPATH
-    cd comps/embeddings/multimodal/multimodal_langchain/
+    cd comps/embeddings/native/multimodal/bridgetower/wrapper/
     docker compose -f docker_compose_multimodal_embedding.yaml up -d
     sleep 2m
 }
